refactor(frontend): decouple contexts logic from hooks logic
