+++ frontend/src/contexts/permission.context.tsx
@@ -1,26 +1,11 @@
-/*
- * Copyright © 2024 Hexastack. All rights reserved.
- *
- * Licensed under the GNU Affero General Public License v3.0 (AGPLv3) with the following additional terms:
- * 1. The name "Hexabot" is a trademark of Hexastack. You may not use this name in derivative works without express written permission.
- * 2. All derivative works must include clear attribution to the original creator and software, Hexastack and Hexabot, in a prominent location (e.g., in the software's "About" section, documentation, and README file).
- */
-
-import {
-  createContext,
-  ReactNode,
-  useCallback,
-  useContext,
-  useMemo,
-} from "react";
+import { createContext, ReactNode, useCallback, useMemo } from "react";
 
 import { Progress } from "@/app-components/displays/Progress";
+import { useUserPermissions } from "@/hooks/entities/auth-hooks";
 import { EntityType } from "@/services/types";
 import { PermissionAction } from "@/types/permission.types";
 
-import { useUserPermissions } from "./entities/auth-hooks";
-
-const PermissionContext = createContext<{
+export const PermissionContext = createContext<{
   getAllowedActions: (_type: EntityType) => undefined | PermissionAction[];
 }>({
   getAllowedActions: (_type: EntityType) => undefined,
@@ -67,18 +52,4 @@ export const PermissionProvider = ({
       {children}
     </PermissionContext.Provider>
   );
-};
-
-export const useHasPermission = () => {
-  const { getAllowedActions } = useContext(PermissionContext);
-  const hasPermission = useCallback(
-    (type: EntityType, action: PermissionAction) => {
-      const allowedActions = getAllowedActions(type);
-
-      return allowedActions?.includes(action);
-    },
-    [getAllowedActions],
-  );
-
-  return hasPermission;
 };
