+++ frontend/src/contexts/setting.context.tsx
@@ -1,0 +1,33 @@
+import { createContext, ReactNode } from "react";
+
+import { Progress } from "@/app-components/displays/Progress";
+import { useLoadSettings } from "@/hooks/entities/auth-hooks";
+import { ISetting } from "@/types/setting.types";
+
+export const SettingsContext = createContext<{
+  settings: { [key: string]: ISetting[] } | undefined;
+}>({ settings: undefined });
+
+SettingsContext.displayName = "SettingsContext";
+
+interface SettingsProviderProps {
+  children: ReactNode;
+}
+
+export const SettingsProvider = ({
+  children,
+}: SettingsProviderProps): JSX.Element => {
+  const { data, isLoading } = useLoadSettings();
+
+  if (isLoading) return <Progress />;
+
+  return (
+    <SettingsContext.Provider
+      value={{
+        settings: data,
+      }}
+    >
+      {children}
+    </SettingsContext.Provider>
+  );
+};
